requirements.txt
numpy==1.22.4
smplx
wandb
trimesh
Cython==0.29.20
chumpy==0.70
hydra-core==1.3.2
Rtree==1.0.1
tqdm
opencv-python==4.7.0.72
open3d==0.17.0
usd-core==22.5
scikit-image==0.19.3
PyYAML==5.4.1
contourpy==1.1.0
torchmetrics==0.6.0
fvcore
iopath